fix: reduce in fp64 in RMSNorm.forward_native to stabilize across torch versions
torch 2.9.1 → 2.11.0 changed the CUDA summation order for `.mean(dim=-1)` /
`.sum(dim=-1)` on fp32 tensors (1-ULP delta). UMT5 keeps hidden states in
fp32, and `RMSNorm.forward_cuda` routes the fp32 + no-residual case to
`forward_native`, so this 1-ULP delta lands in the variance computation,
propagates through 24 UMT5 encoder blocks into DiT cross-attention via
`encoder_hidden_states`, and compounds over 50 UniPC denoise steps into a
multi-percent rel_err on the final latent (verified on wan2_1_t2v_1.3b
and wan2_1_i2v_14b_480P_2gpu — see specs/torch-211-bisection/findings.md).

Upcasting `.mean(dim=-1)` to fp64 for the reduction — independent of the
torch reduction kernel — produces byte-identical denoise output across
all 50 steps and ~95M fp32 values between torch 2.9.1 and 2.11.0 in both
the 1.3B 1-GPU t2v case and the 14B 2-GPU ulysses i2v case (level-a step
diff: 102/102 step-rank pairs L2 = 0.0 with the patch enabled).

fp64 multiplies on Hopper run at ~1/64 fp32 throughput, but this is one
reduction per RMSNorm per step over (B, seq_len, hidden) — negligible
vs the surrounding fp32/bf16 matmuls.

+++ python/sglang/multimodal_gen/runtime/layers/layernorm.py
@@ -142,7 +142,14 @@ class RMSNorm(CustomOp):
 
             x_var = x[..., : self.variance_size_override]
 
-        variance = x_var.pow(2).mean(dim=-1, keepdim=True)
+        # Reduce in fp64 to be invariant to torch's CUDA reduction-kernel changes
+        # (fp32 .mean(dim=-1) summation order changed between torch 2.9.1 and
+        # 2.11.0, producing a 1-ULP variance delta that compounds through UMT5
+        # encoder blocks and across 50 denoise steps to a multi-percent latent
+        # drift in diffusion pipelines). fp64 mul throughput on Hopper is ~1/64
+        # of fp32, but this runs once per RMSNorm per step on a (B, seq, hidden)
+        # reduction — negligible vs the surrounding matmuls.
+        variance = x_var.double().pow(2).mean(dim=-1, keepdim=True).float()
         x = x * torch.rsqrt(variance + self.variance_epsilon)
         x = (x * self.weight).to(orig_dtype)
         if residual is None:
